refactor: split page into components
